pyproject.toml
[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[project]
name = "mcp-common"
version = "0.9.4"
description = "Oneiric-Native foundation library providing battle-tested patterns for MCP (Model Context Protocol) servers, with YAML configuration, Rich UI, and CLI lifecycle management"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    { name = "Les Leslie", email = "les@wedgwoodwebworks.com" },
]
keywords = [
    "mcp",
    "model-context-protocol",
    "utilities",
    "http-client",
    "configuration",
    "security",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Typing :: Typed",
]
dependencies = [
    "oneiric>=0.3.6",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.0",
    "psutil>=7.2.1",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "typer>=0.21.0",
    "PyJWT>=2.8.0",
    "cryptography>=41.0.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
# Prompting/notification backends
macos-prompts = [
    "pyobjc-core>=10.0",
    "pyobjc-framework-Cocoa>=10.0",
]
terminal-prompts = [
    "prompt-toolkit>=3.0",
]
all-prompts = [
    "mcp-common[macos-prompts,terminal-prompts]",
]

# Tree-sitter parsing support
treesitter = [
    "tree-sitter>=0.24.0",
    "tree-sitter-python>=0.23.0",
    "tree-sitter-go>=0.23.0",
]

# Development dependencies
dev = [
    "crackerjack>=0.47.0",
    "pytest-benchmark>=4.0.0",
    "uv-bump>=0.4.0",
]

# All optional features
all = [
    "mcp-common[all-prompts,dev,treesitter]",
]

[project.license]
text = "BSD-3-Clause"

[project.urls]
Homepage = "https://github.com/lesleslie/mcp-common"
Documentation = "https://github.com/lesleslie/mcp-common#readme"
Repository = "https://github.com/lesleslie/mcp-common"
Issues = "https://github.com/lesleslie/mcp-common/issues"

[tool.hatchling.build.targets.wheel]
packages = [
    "mcp_common",
]

[tool.ruff]
target-version = "py313"
line-length = 88
exclude = [
    "tests/",
    "test_*.py",
    "*_test.py",
]

[tool.ruff.lint]
extend-select = [
    "C901",
    "F",
    "I",
    "UP",
]
ignore = [
    "E402",
    "F821",
]
fixable = [
    "ALL",
]

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.pytest.ini_options]
asyncio_mode = "auto"
timeout = 600
testpaths = [
    "tests",
]
addopts = [
    "--cov=mcp-common",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html:htmlcov",
    "--cov-report=json",
    "--strict-markers",
    "--strict-config",
    "-ra",
    "--tb=short",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "security: Security-focused tests",
    "performance: Performance and load tests",
    "slow: Slow tests (>2s)",
    "benchmark: Benchmark tests (deferred)",
    "smoke: Quick smoke tests",
    "regression: Regression tests",
    "api: API endpoint tests",
]

[tool.coverage.run]
branch = true
parallel = true
concurrency = [
    "multiprocessing",
]
source = [
    "mcp-common",
]
data_file = ".coverage"
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/__init__.py",
    "*/_version.py",
    "*/conftest.py",
]

[tool.coverage.report]
exclude_also = [
    "pragma: no cover",
    "def __repr__",
    "def __str__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "@abstractmethod",
    "pass",
    "raise ImportError",
    "except ImportError",
]

[tool.pyright]
include = [
    "mcp-common",
]
exclude = [
    "tests/*",
    "scratch",
    ".venv",
    "build",
    "dist",
]
typeCheckingMode = "strict"
pythonVersion = "3.13"

[tool.bandit]
target = [
    "mcp-common",
]
skips = [
    "B101",
    "B110",
    "B112",
    "B404",
    "B603",
    "B607",
]
exclude_dirs = [
    "tests/",
]

[tool.codespell]
quiet-level = 3
ignore-words-list = "crate,uptodate,nd,nin"

[tool.creosote]
paths = [
    "mcp-common",
]
deps-file = "pyproject.toml"
exclude-deps = [
    "hatchling",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-mock",
    "pytest-xdist",
    "pytest-timeout",
    "pytest-benchmark",
    "ruff",
    "pyright",
    "bandit",
    "creosote",
    "codespell",
    "refurb",
    "complexipy",
]

[tool.refurb]
enable_all = false
quiet = true
python_version = "3.13"

[tool.complexipy]
default_pattern = "**/*.py"
exclude_patterns = [
    "**/tests/**",
    "**/test_*.py",
]
max_complexity = 15

[dependency-groups]
dev = [
    "crackerjack>=0.49.8",
]
